+++ libs/model/src/utils/index.ts
@@ -1,5 +1,6 @@
 export * from './denormalizedCountUtils';
 export * from './getDelta';
+export * from './parseUserMentions';
 export * from './sanitizeQuillText';
 export * from './updateTags';
 export * from './utils';
