+++ packages/commonwealth/client/scripts/views/components/react_quill_editor/render_quill_delta.tsx
@@ -1,7 +1,6 @@
-import React, { ReactElement } from 'react';
-
+import { preprocessQuillDeltaForRendering } from '@hicommonwealth/shared';
 import { loadScript } from 'helpers';
-import { preprocessQuillDeltaForRendering } from '../../../../../shared/utils';
+import React, { ReactElement } from 'react';
 
 type TempList = Array<
   Array<{
