+++ packages/commonwealth/server/controllers/server_comments_methods/update_comment.ts
@@ -3,16 +3,17 @@ import {
   AddressInstance,
   CommentAttributes,
   UserInstance,
-} from '@hicommonwealth/model';
-import { NotificationCategories, ProposalType } from '@hicommonwealth/shared';
-import { WhereOptions } from 'sequelize';
-import { validateOwner } from 'server/util/validateOwner';
-import { renderQuillDeltaToText } from '../../../shared/utils';
-import {
   emitMentions,
   findMentionDiff,
   parseUserMentions,
-} from '../../util/parseUserMentions';
+} from '@hicommonwealth/model';
+import {
+  NotificationCategories,
+  ProposalType,
+  renderQuillDeltaToText,
+} from '@hicommonwealth/shared';
+import { WhereOptions } from 'sequelize';
+import { validateOwner } from 'server/util/validateOwner';
 import { addVersionHistory } from '../../util/versioning';
 import { ServerCommentsController } from '../server_comments_controller';
 import { EmitOptions } from '../server_notifications_methods/emit';
