+++ packages/commonwealth/server/controllers/server_threads_methods/update_thread.ts
@@ -7,8 +7,15 @@ import {
   ThreadAttributes,
   ThreadInstance,
   UserInstance,
+  emitMentions,
+  findMentionDiff,
+  parseUserMentions,
 } from '@hicommonwealth/model';
-import { NotificationCategories, ProposalType } from '@hicommonwealth/shared';
+import {
+  NotificationCategories,
+  ProposalType,
+  renderQuillDeltaToText,
+} from '@hicommonwealth/shared';
 import _ from 'lodash';
 import {
   Op,
@@ -18,12 +25,7 @@ import {
   WhereOptions,
 } from 'sequelize';
 import { MixpanelCommunityInteractionEvent } from '../../../shared/analytics/types';
-import { renderQuillDeltaToText, validURL } from '../../../shared/utils';
-import {
-  emitMentions,
-  findMentionDiff,
-  parseUserMentions,
-} from '../../util/parseUserMentions';
+import { validURL } from '../../../shared/utils';
 import { findAllRoles } from '../../util/roles';
 import { addVersionHistory } from '../../util/versioning';
 import { TrackOptions } from '../server_analytics_controller';
